+++ GraphingWithDADMIN.py
@@ -51,41 +51,12 @@ for (K, L) in StatementType_Keylist.items():
 # TODO: set height of Listbox to the length of the list or ensure widow height is enough
 
 
-# 'keyevent' gets passed implicitly
-def KeybindTest(keyevent):
-    print("bound function called")
-    print(f"passed: {keyevent}")
 
 
-tkWindow.bind('<Key-Return>', KeybindTest)
 
 
-def MakeButtons(count):
-    queue = []
-    for x in range(count):
-        newcounter = tkinter.IntVar(value=x)
-        def newfun(somecounter=newcounter): somecounter.set(somecounter.get() + 1)
-        newbutton = tkinter.Button(tkWindow, textvariable=newcounter, anchor=tkinter.NE, command=newfun)
-        newbutton.pack(side=tkinter.LEFT, before=frame)
-        queue.append((newbutton, newcounter))
-    return queue
 
 
-def ButtonExperiment():
-    firstset, newbuttonlist = MakeButtons(3), MakeButtons(3)
-    for (FSB, FSC) in firstset:
-        FSB.pack(side=tkinter.TOP, anchor=tkinter.SW, before=frame)
-        FSC.set(FSC.get() + 1)
-        FSB["height"] += FSC.get()
-        FSB["width"] += FSC.get()
-    for (newbuttons, counters) in [*firstset, *newbuttonlist]:
-        def switcheroo(NB=newbuttons, NC=counters):
-            NC.set(NC.get() + 1)
-            NB["width"] = NC.get()
-            NB["height"] = NC.get()
-        newbuttons["command"] = switcheroo
-    print(f"tkWindow \n {tkWindow.keys()}\n")
-    print(f"frame \n {frame.keys()}\n")
 
 
 # Create a custom formatter to format the y-axis labels
@@ -141,11 +112,5 @@ def PlotWantedKeys(thejson, figure=None):
 
 
 if __name__ == '__main__':
-    #ButtonExperiment()
     tkWindow.mainloop()
-    #userselection = LoadUserSelections()
-    #newfig = None
-    #for F in userselection:
-    #    newfig = PlotWantedKeys(F)
-    #pyplot.show()
     print("window closed")
